kgsl: Obtain ddr type using smem
Currently, ddr type is being obtained by parsing device tree.
Instead read ddr type using smem as device tree node is being
deprecated.

(cherry picked from commit ad6ac9ef5bc7b7c6b4e2467a77600b93ede6198f
https://git.codelinaro.org/clo/le/platform/vendor/qcom/opensource/graphics-kernel.git gfx-kernel.le.0.0.r1-rel)

+++ ubuntu/qcom/graphics/kgsl_util.h
@@ -264,10 +264,9 @@ static inline int qcom_scm_kgsl_set_smmu_lpac_aperture(
 /**
  * kgsl_get_ddrtype - Type of ddr on the current device
  *
- * Return: non-zero positive value on success or negative
- * error code cast to u64 on failure.
+ * Return: ddr type on success or negative error on failure.
  */
-u64 kgsl_get_ddrtype(void);
+int kgsl_get_ddrtype(void);
 
 #if !IS_ENABLED(CONFIG_QCOM_SCM_ADDON)
 static inline int qcom_scm_kgsl_init_regs(u32 gpu_req)
